fix: Implement HLS resume with stream offset tracking for accurate scrubber positioning
This commit fixes video resume functionality by implementing a comprehensive solution
for HLS progressive transcoding with the ?start= parameter.

**Problem:**
Videos with saved progress would start from the beginning instead of resuming from
the saved position. When using on-demand HLS transcoding, the backend transcodes from
the beginning, making the resume position unseekable.

**Root Cause:**
1. Resume logic was checking `is_watched` flag (90%+) instead of actual completion
2. HLS streams couldn't seek to positions not yet transcoded
3. Stream offset wasn't tracked when backend started transcoding from resume position

**Solution:**
1. **Resume Logic Fix**: Check actual completion (within 1s of end) instead of is_watched flag
2. **Backend Start Parameter**: Add ?start= parameter to HLS manifest URL to tell backend
   to start transcoding from resume position
3. **Stream Offset Tracking**: Track streamOffsetRef to map video.currentTime (stream timeline)
   to actual position in full video for correct scrubber display
4. **HLS Detection Fix**: Use .includes('.m3u8') instead of .endsWith('.m3u8') to handle
   query parameters in manifest URLs
5. **Offset Preservation**: Only set streamOffset on initial load (when === 0) to prevent
   seek operations from being overwritten on manifest reload

**Technical Details:**
- When resuming from position X: Backend transcodes from X, HLS stream timeline starts at 0
- video.currentTime reports position in stream (0-N), not in full video (X-duration)
- streamOffsetRef.current = X allows: displayTime = video.currentTime + streamOffsetRef.current
- Seeking converts: streamTime = targetTime - streamOffsetRef.current
- Large seeks (>30s) reload manifest with new ?start= and update offset

**Files Changed:**
- web/src/lib/hooks/useMediaPlayback.ts: Resume logic, ?start= parameter
- web/src/components/media/VideoPlayer/VideoPlayer.tsx: Stream offset tracking, seek handling
- web/src/components/media/VideoPlayerContainer/VideoPlayerContainer.tsx: Cleaned up debug logs

Fixes video resume functionality for HLS progressive transcoding.

+++ internal/application/progress/dto.go
@@ -8,10 +8,10 @@ import (
 
 // UpdateProgressRequest represents a request to update watch progress.
 type UpdateProgressRequest struct {
-	MediaID         int64 `json:"media_id"`
-	UserID          int64 `json:"user_id"`
-	ProgressSeconds int   `json:"progress_seconds"`
-	DurationSeconds int   `json:"duration_seconds"`
+	MediaID         int64   `json:"media_id"`
+	UserID          int64   `json:"user_id"`
+	ProgressSeconds float64 `json:"progress_seconds"`
+	DurationSeconds float64 `json:"duration_seconds"`
 }
 
 // WatchProgressResponse represents watch progress information in API responses.
@@ -19,8 +19,8 @@ type WatchProgressResponse struct {
 	ID                 int64     `json:"id"`
 	MediaID            int64     `json:"media_id"`
 	UserID             int64     `json:"user_id"`
-	ProgressSeconds    int       `json:"progress_seconds"`
-	DurationSeconds    int       `json:"duration_seconds"`
+	ProgressSeconds    float64   `json:"progress_seconds"`
+	DurationSeconds    float64   `json:"duration_seconds"`
 	ProgressPercentage float64   `json:"progress_percentage"`
 	IsWatched          bool      `json:"is_watched"`
 	LastWatchedAt      time.Time `json:"last_watched_at"`
